Create Project 2: String Operations
using string operation like: indexing, traversing, slicing, searching, comparison, comparison for ordering to manipulate strings to solve word problems. 

Question 1: You have a hunch students going to schools ending in a vowel are smarter than students who go to a school ending in a consonant. Write a code that will tell user if a school will foster smart students or not.

Question 2: Given a list of of first name and academic grades, user will have to alter a 'c' grade to an 'a'. This code will also tell us the position where this 'c' grade is located at.