Create Project 3: Reading and Writing Files
Q1: Given a dataset, I have to create a code that counts the amount of times 'Syracuse' and 'Rochester' appears at the end of the sentences. The final output will be the amount of times printed out. 

Q2: Main focus of question 2 is to use data normalization. The data is from a dataset examining food desserts. The percent of households located in a food dessert will be recalculated  for a census block  group.